rm duplicate memory_checking
