+++ src/lasso/mod.rs
@@ -1,4 +1,3 @@
 pub mod densified;
 pub mod memory_checking;
-pub mod memory_checking_updated;
 pub mod surge;
