+++ src/lasso/surge.rs
@@ -24,6 +24,8 @@ use ark_serialize::*;
 use ark_std::log2;
 use merlin::Transcript;
 
+use super::memory_checking::HashLayerProof;
+
 pub struct SparsePolyCommitmentGens<G> {
   pub gens_combined_l_variate: PolyCommitmentGens<G>,
   pub gens_combined_log_m_variate: PolyCommitmentGens<G>,
@@ -99,7 +101,7 @@ pub struct PrimarySumcheck<G: CurveGroup> {
 pub struct SparsePolynomialEvaluationProof<G: CurveGroup, S: JoltStrategy<G::ScalarField>> {
   comm_derefs: CombinedTableCommitment<G>,
   primary_sumcheck: PrimarySumcheck<G>,
-  memory_check: MemoryCheckingProof<G>,
+  memory_check: MemoryCheckingProof<G, HashLayerProof<G>>,
   sparsity: usize,
   memory_size: usize,
   _marker: PhantomData<S>,
